Rename scheduling mod to schedule. Suppress "module inception" clippy lint.
