+++ benches/benches/bevy_ecs/scheduling/running_systems.rs
@@ -1,4 +1,4 @@
-use bevy_ecs::{component::Component, scheduling::Schedule, system::Query, world::World};
+use bevy_ecs::{component::Component, schedule::Schedule, system::Query, world::World};
 use criterion::Criterion;
 
 #[derive(Component)]
