+++ crates/bevy_diagnostic/src/lib.rs
@@ -5,7 +5,7 @@ mod log_diagnostics_plugin;
 mod system_information_diagnostics_plugin;
 
 use bevy_app::prelude::*;
-use bevy_ecs::scheduling::IntoSystemConfig;
+use bevy_ecs::schedule::IntoSystemConfig;
 pub use diagnostic::*;
 pub use entity_count_diagnostics_plugin::EntityCountDiagnosticsPlugin;
 pub use frame_time_diagnostics_plugin::FrameTimeDiagnosticsPlugin;
