+++ crates/bevy_ecs/src/system/exclusive_function_system.rs
@@ -154,8 +154,8 @@ where
         );
     }
 
-    fn default_system_sets(&self) -> Vec<Box<dyn crate::scheduling::SystemSet>> {
-        let set = crate::scheduling::SystemTypeSet::<F>::new();
+    fn default_system_sets(&self) -> Vec<Box<dyn crate::schedule::SystemSet>> {
+        let set = crate::schedule::SystemTypeSet::<F>::new();
         vec![Box::new(set)]
     }
 }
