+++ crates/bevy_ecs/src/system/system.rs
@@ -65,7 +65,7 @@ pub trait System: Send + Sync + 'static {
     fn update_archetype_component_access(&mut self, world: &World);
     fn check_change_tick(&mut self, change_tick: u32);
     /// Returns the system's default [system sets](crate::schedule::SystemSet).
-    fn default_system_sets(&self) -> Vec<Box<dyn crate::scheduling::SystemSet>> {
+    fn default_system_sets(&self) -> Vec<Box<dyn crate::schedule::SystemSet>> {
         Vec::new()
     }
     /// Gets the system's last change tick
