+++ crates/bevy_render/src/pipelined_rendering.rs
@@ -2,7 +2,7 @@ use async_channel::{Receiver, Sender};
 
 use bevy_app::{App, AppLabel, CoreSchedule, Plugin, SubApp};
 use bevy_ecs::{
-    scheduling::{MainThreadExecutor, Schedule},
+    schedule::{MainThreadExecutor, Schedule},
     system::Resource,
     world::{Mut, World},
 };
