+++ examples/games/alien_cake_addict.rs
@@ -2,7 +2,7 @@
 
 use std::f32::consts::PI;
 
-use bevy::{ecs::scheduling::SystemSet, prelude::*};
+use bevy::{ecs::schedule::SystemSet, prelude::*};
 use rand::Rng;
 
 #[derive(Clone, Eq, PartialEq, Debug, Hash)]
