+++ tools/ci/src/main.rs
@@ -16,8 +16,9 @@ bitflags! {
     }
 }
 
-const CLIPPY_FLAGS: [&str; 8] = [
+const CLIPPY_FLAGS: &[&str] = &[
     "-Aclippy::type_complexity",
+    "-Aclippy::module_inception",
     "-Wclippy::doc_markdown",
     "-Wclippy::redundant_else",
     "-Wclippy::match_same_arms",
